Run a full workflow.

This is the main function of zoon. The arguments should specify at least five modules, at least one of each type. If modules do not have any arguments to be specific (or defaults are being used) then simply give the names of the module. If arguments are needed give the modules in the form of a function e.g. occurrence = AModule(para1 = 2, para2 = 'detail')

workflow(occurrence, covariate, process, model, output,
  forceReproducible = FALSE)

Arguments

occurrence

Occurrence module to be used.

covariate

Covariate module to be used.

process

Process module to be used.

model

SDM model module to be used.

output

Output module to be used.

forceReproducible

Logical whether to force zoon to collect modules from the online repo. This ensure the analysis is reproducible.

Value

A list with the results of each module and a copy of the code used to execute the workflow. If the workflow fails a partial list is saved to a temporary file for debugging.

Examples

# run a workflow, using the logistic regression model
## Not run: ------------------------------------
# 
# work1 <- workflow(occurrence = UKAnophelesPlumbeus,
#                 covariate = UKAir,
#                 process = Background(n = 70),
#                 model = LogisticRegression,
#                 output = SameTimePlaceMap)
# 
# str(work1, 1)
# 
# work2 <- workflow(UKAnophelesPlumbeus,
#                  UKAir,
#                  OneHundredBackground,   
#                  RandomForest,
#                  PrintMap)
# 
## ---------------------------------------------